requirements.txt
llama-index==0.12.8
llama-index-vector-stores-chroma==0.4.1
llama-index-embeddings-openai==0.3.1
llama-index-llms-openai==0.3.3
chromadb==0.6.3
pypdf==5.4.0
youtube-transcript-api==1.2.4
streamlit==1.45.0
python-dotenv==1.1.0
requests==2.32.3
pytest==8.3.5
